Move around test vectors, impl From<SpendingKey> for FullViewingKey
