+++ zebra-chain/src/orchard/address.rs
@@ -61,7 +61,7 @@ mod tests {
 
         let spending_key = keys::SpendingKey::new(&mut OsRng, network);
 
-        let full_viewing_key = keys::FullViewingKey::from_spending_key(spending_key);
+        let full_viewing_key = keys::FullViewingKey::from(spending_key);
 
         // Default diversifier, where index = 0.
         let diversifier_key = keys::DiversifierKey::from(full_viewing_key);
